Split District to District, DistrictComparator, DistrictEnums. ColonyActions.
