+++ sources/Colonization/AngelScriptBinders/Core/District/BindColonyActions.hpp
@@ -1,0 +1,6 @@
+#pragma once
+#include <Urho3D/AngelScript/Script.h>
+namespace Colonization
+{
+void BindColonyActions (Urho3D::Script *script);
+}
