+++ sources/Colonization/AngelScriptBinders/Core/District/BindDistrictEnums.hpp
@@ -1,0 +1,6 @@
+#pragma once
+#include <Urho3D/AngelScript/Script.h>
+namespace Colonization
+{
+void BindDistrictEnums (Urho3D::Script *script);
+}
