+++ sources/Colonization/Backend/ColoniesActionsProcessor.cpp
@@ -6,6 +6,7 @@
 #include <Urho3D/Scene/Scene.h>
 
 #include <cmath>
+#include <Colonization/Core/District/ColonyActions.hpp>
 #include <Colonization/Backend/PlayersManager.hpp>
 #include <Colonization/Utils/Network/NetworkUpdateCounter.hpp>
 #include <Colonization/Utils/Serialization/Categories.hpp>
