+++ sources/Colonization/Core/District/District.hpp
@@ -4,49 +4,11 @@
 #include <Urho3D/Scene/Node.h>
 #include <Urho3D/Container/Str.h>
 #include <Urho3D/Scene/LogicComponent.h>
+#include <Colonization/Core/District/DistrictEnums.hpp>
 
 namespace Colonization
 {
-// TODO: Maybe split to District, DistrictComparator, ColonyActions, DistrictEnums.
 class Map;
-class District;
-typedef bool (*DistrictComparator) (const District *first, const District *second);
-namespace DistrictComparators
-{
-bool HigherLogistics (const District *first, const District *second);
-}
-
-enum ClimateType
-{
-    CLIMATE_TROPICAL = 0,
-    CLIMATE_HOT = 1,
-    CLIMATE_TEMPERATE = 2,
-    CLIMATE_TEMPERATE_CONTINENTAL = 3,
-    CLIMATE_DESERT = 4,
-    CLIMATE_COLD = 5,
-};
-
-enum NativesCharacterType
-{
-    NATIVES_CHARACTER_FRIENDLY = 0,
-    NATIVES_CHARACTER_MEDIUM = 1,
-    NATIVES_CHARACTER_COLD = 2,
-    NATIVES_CHARACTER_ISOLATIONIST = 3,
-    NATIVES_CHARATER_AGGRESSIVE = 4
-};
-
-const Urho3D::StringHash COLONY_ACTION_ID ("ColonyActionId");
-const Urho3D::StringHash COLONY_ACTION_PROGRESS ("ColonyActionProgress");
-namespace ColonyActions
-{
-const Urho3D::StringHash BUILD_WAR_SHIP ("BuildWarShip");
-namespace BuildWarShip
-{
-/// District hash (sea district, neighbor of colony).
-const Urho3D::StringHash TARGET_DISTRICT ("TargetDistrict");
-}
-}
-
 class District : public Urho3D::Component
 {
 URHO3D_OBJECT (District, Component)
