+++ sources/Colonization/Core/District/DistrictEnums.hpp
@@ -1,0 +1,22 @@
+#pragma once
+namespace Colonization
+{
+enum ClimateType
+{
+    CLIMATE_TROPICAL = 0,
+    CLIMATE_HOT = 1,
+    CLIMATE_TEMPERATE = 2,
+    CLIMATE_TEMPERATE_CONTINENTAL = 3,
+    CLIMATE_DESERT = 4,
+    CLIMATE_COLD = 5,
+};
+
+enum NativesCharacterType
+{
+    NATIVES_CHARACTER_FRIENDLY = 0,
+    NATIVES_CHARACTER_MEDIUM = 1,
+    NATIVES_CHARACTER_COLD = 2,
+    NATIVES_CHARACTER_ISOLATIONIST = 3,
+    NATIVES_CHARATER_AGGRESSIVE = 4
+};
+}
