+++ sources/Colonization/Core/Map.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include <Urho3D/Scene/Node.h>
 #include <Urho3D/Container/Vector.h>
-#include <Colonization/Core/District.hpp>
+#include <Colonization/Core/District/District.hpp>
 
 namespace Colonization
 {
