+++ sources/Tests/TestMapMaskGeneration.cpp
@@ -6,7 +6,7 @@
 #include <Urho3D/Scene/SceneEvents.h>
 #include <Urho3D/IO/Log.h>
 
-#include <Colonization/Core/District.hpp>
+#include <Colonization/Core/District/District.hpp>
 #include <Colonization/Core/Map.hpp>
 #include <Colonization/Core/GameConfiguration.hpp>
 #include <Colonization/Core/Unit.hpp>
