+++ sources/Tests/TestMapPathFinding.cpp
@@ -4,7 +4,7 @@
 #include <Urho3D/Scene/Scene.h>
 #include <Urho3D/IO/Log.h>
 
-#include <Colonization/Core/District.hpp>
+#include <Colonization/Core/District/District.hpp>
 #include <Colonization/Core/Map.hpp>
 #include <Colonization/Backend/UnitsManager.hpp>
 #include <Colonization/Core/GameConfiguration.hpp>
